* develop: (36 commits)
  Fix heroic build script
  fix reds connection
  Fix import typo.
  Fix test script.
  Add dependency service.
  Remove yarn.
  Complete receiving update from FB.
  Close Redis connect after action done.
  Add instruction in README.
  Create Redis client with NODE_ENV or config file.
  Add schema.
  Add live relatived GraphQL resolvers.
  Remove code for testing.
  Attach socket.to to express.
  Complete socket server.
  Add socket unit test.
  Add socket and reds dependency.
  Add express-x-hub and authenticate FB subscribe.
  GraphQL authorization.
  Allow path to facebook
  ...

+++ src/data/mutations/deleteAdmin.js
@@ -1,6 +1,6 @@
 import { GraphQLInt, GraphQLNonNull, GraphQLBoolean } from 'graphql';
 
-import Admin from '../models/admin';
+import { Admin } from '../models';
 
 const deleteAdmin = {
   type: GraphQLBoolean,
@@ -8,7 +8,11 @@ const deleteAdmin = {
     id: { type: new GraphQLNonNull(GraphQLInt) },
   },
 
-  async resolve(obj, {id}) {
+  async resolve(obj, {id}, context, { rootValue: { request } }) {
+    if (!request.user) {
+      return false;
+    }
+
     const affectedRows = Admin.destroy({ where: { id: id } });
 
     return (affectedRows <= 0 ? false : true);
